+++ mrp_multi_level/readme/HISTORY.rst
@@ -3,6 +3,9 @@
 
 * [FIX] User and system locales doesn't break MRP calculation.
   (`#290 <https://github.com/OCA/manufacture/pull/290>`_)
+* [FIX] Working Hours are now defined only at Warehouse level and displayed
+  as a related on MRP Areas.
+  (`#290 <https://github.com/OCA/manufacture/pull/290>`__)
 
 11.0.1.0.0 (2018-07-09)
 ~~~~~~~~~~~~~~~~~~~~~~~